revert elem size check in create

+++ vector.h
@@ -12,7 +12,7 @@ typedef struct Vector {
 	size_t _elem_size;
 } Vector;
 
-Vector* vector_create(int);
+Vector* vector_create(size_t);
 void vector_delete(Vector*);
 
 void* vector_get(Vector*, size_t);
